requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.2
flask-cors==4.0.0
